+++ src/i18n/format.rs
@@ -1,3 +1,4 @@
 pub const FORMAT_SHARE_DETAILS: &str = "format-share-details";
+pub const FORMAT_EXPENSE_DETAILS: &str = "format-expense-details";
 pub const FORMAT_EXPENSE: &str = "format-expense";
 pub const FORMAT_TRANSFER: &str = "format-transfer";